# docs/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: LeetGaming Pro API
  description: |
    # 🎮 LeetGaming Pro - Competitive Gaming Platform API

    Welcome to the LeetGaming Pro API documentation. This API powers a professional-grade
    competitive gaming platform with features including:

    - **Authentication & Security**: OAuth2, MFA/TOTP, JWT tokens
    - **Matchmaking**: Glicko-2 rating system, skill-based matching
    - **Tournaments**: Bracket generation, prize pools, blockchain verification
    - **Replays**: Upload, analysis, sharing with demo parsing
    - **Squads**: Team management, invitations, statistics
    - **Billing**: Subscriptions, withdrawals, audit trails

    ## 🔐 Authentication

    Most endpoints require authentication via Bearer token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## 📊 Rate Limits

    | Tier | Requests/min | Burst |
    |------|--------------|-------|
    | Free | 60 | 10 |
    | Pro | 300 | 50 |
    | Enterprise | 1000 | 200 |

    ## 🌐 Base URLs

    | Environment | URL |
    |-------------|-----|
    | Production | https://api.leetgaming.gg |
    | Staging | https://api.staging.leetgaming.gg |
    | Development | http://localhost:8080 |

  version: 2.0.0
  contact:
    name: LeetGaming Pro Support
    url: https://leetgaming.gg/support
    email: api@leetgaming.gg
  license:
    name: Proprietary
    url: https://leetgaming.gg/terms
servers:
  - url: https://api.leetgaming.gg
    description: Production
...
    description: Staging
...
paths:
...
  /auth/mfa/setup:
    post:
      tags: [Authentication]
      summary: Set up MFA/TOTP
      operationId: setupMFA
      security:
        - bearerAuth: []
      responses:
        "200":
          description: MFA setup initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFASetupResponse"
  /auth/mfa/verify:
    post:
      tags: [Authentication]
      summary: Verify MFA code
      operationId: verifyMFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        "200":
          description: MFA verified
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Players
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    MFASetupResponse:
      type: object
      properties:
        secret:
          type: string
        qr_code_url:
          type: string
          format: uri
        recovery_codes:
          type: array
          items:
            type: string
...
security:
  - bearerAuth: []